% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_model.R
\name{add_edges}
\alias{add_edges}
\title{Add edges to a VAST model}
\usage{
add_edges(model, ...)
}
\arguments{
\item{model}{A \code{vast_model} object.}

\item{...}{One or more \code{vast_edge} objects, or a single list of them.}
}
\value{
The updated \code{vast_model}.
}
\description{
Add edges to a VAST model
}
\examples{
m <- vast_model() |>
  add_nodes(vast_concept("X"), vast_concept("Y")) |>
  add_edges(vast_causation("X", "Y", strength = 0.5))

}
