% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_model.R
\name{add_nodes}
\alias{add_nodes}
\title{Add nodes to a VAST model}
\usage{
add_nodes(model, ...)
}
\arguments{
\item{model}{A \code{vast_model} object.}

\item{...}{One or more \code{vast_node} objects, or a single list of them.}
}
\value{
The updated \code{vast_model} (invisibly pipe-friendly).
}
\description{
Add nodes to a VAST model
}
\examples{
m <- vast_model() |>
  add_nodes(
    vast_concept("X"),
    vast_concept("Y")
  )

}
