% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_model.R
\name{.edge_to_dot}
\alias{.edge_to_dot}
\title{Generate DOT for a single edge}
\usage{
.edge_to_dot(edge, fontsize_override = NULL)
}
\arguments{
\item{edge}{A vast_edge object.}

\item{fontsize_override}{Optional fontsize for the edge label.}
}
\description{
Generate DOT for a single edge
}
\keyword{internal}
