% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_model.R
\name{.group_to_dot}
\alias{.group_to_dot}
\title{Generate DOT for a group (subgraph cluster)}
\usage{
.group_to_dot(
  group,
  all_groups,
  all_nodes,
  indent = "  ",
  label_overrides = NULL,
  skip_node_ids = NULL,
  fontsize_nodes = NULL,
  fontsize_names = NULL
)
}
\arguments{
\item{group}{A vast_group object.}

\item{all_groups}{Named list of all groups in the model.}

\item{all_nodes}{Named list of all nodes in the model.}

\item{indent}{Current indentation string.}

\item{label_overrides}{Named character vector of label overrides.}

\item{skip_node_ids}{Character vector of node IDs to suppress.}

\item{fontsize_nodes}{Numeric. Font size override for non-name nodes.}

\item{fontsize_names}{Numeric. Font size override for name nodes.}
}
\description{
Generate DOT for a group (subgraph cluster)
}
\keyword{internal}
