% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_model.R
\name{.naming_legend_to_dot}
\alias{.naming_legend_to_dot}
\title{Generate a naming legend as a compact HTML table node for separated mode}
\usage{
.naming_legend_to_dot(model, naming_info)
}
\description{
Generate a naming legend as a compact HTML table node for separated mode
}
\keyword{internal}
