% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_model.R
\name{.node_to_dot}
\alias{.node_to_dot}
\title{Generate DOT for a single node}
\usage{
.node_to_dot(node, label_override = NULL, fontsize_override = NULL)
}
\arguments{
\item{node}{A vast_node object.}

\item{label_override}{Optional label to display instead of node$label.}

\item{fontsize_override}{Optional fontsize to use instead of node$fontsize.}
}
\description{
Generate DOT for a single node
}
\keyword{internal}
