% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{print.vast_edge}
\alias{print.vast_edge}
\title{Print a VAST edge}
\usage{
\method{print}{vast_edge}(x, ...)
}
\arguments{
\item{x}{A \code{vast_edge} object.}

\item{...}{Ignored.}
}
\description{
Print a VAST edge
}
