% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{print.vast_group}
\alias{print.vast_group}
\title{Print a VAST group}
\usage{
\method{print}{vast_group}(x, ...)
}
\arguments{
\item{x}{A \code{vast_group} object.}

\item{...}{Ignored.}
}
\description{
Print a VAST group
}
