% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_model.R
\name{print.vast_model}
\alias{print.vast_model}
\title{Print a VAST model summary}
\usage{
\method{print}{vast_model}(x, ...)
}
\arguments{
\item{x}{A \code{vast_model} object.}

\item{...}{Ignored.}
}
\description{
Print a VAST model summary
}
