% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{print.vast_node}
\alias{print.vast_node}
\title{Print a VAST node}
\usage{
\method{print}{vast_node}(x, ...)
}
\arguments{
\item{x}{A \code{vast_node} object.}

\item{...}{Ignored.}
}
\description{
Print a VAST node
}
