% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vastR-package.R
\docType{package}
\name{vastR-package}
\alias{vastR}
\alias{vastR-package}
\title{vastR: Visual Argument Structure Tool for R}
\description{
An R implementation of the Visual Argument Structure Tool (VAST)
by Leising, Grenke, & Cramer (2023, Meta-Psychology).

VAST provides a visual language for mapping argument structures. This
package lets you programmatically construct VAST diagrams and render
them via Graphviz / DiagrammeR.
}
\section{Workflow}{

\enumerate{
\item Create a model with \code{\link[=vast_model]{vast_model()}}
\item Add nodes with \code{\link[=add_nodes]{add_nodes()}} using constructors like \code{\link[=vast_concept]{vast_concept()}},
\code{\link[=vast_name]{vast_name()}}, \code{\link[=vast_diamond]{vast_diamond()}}, etc.
\item Add edges with \code{\link[=add_edges]{add_edges()}} using \code{\link[=vast_causation]{vast_causation()}},
\code{\link[=vast_implication]{vast_implication()}}, etc.
\item Optionally add higher-order concept groups with \code{\link[=add_groups]{add_groups()}}
\item Render with \code{\link[=vast_render]{vast_render()}} or export with \code{\link[=vast_export_svg]{vast_export_svg()}}
}
}

\references{
Leising, D., Grenke, O., & Cramer, M. (2023). Visual Argument Structure
Tool (VAST) Version 1.0. \emph{Meta-Psychology}, 7.
\doi{10.15626/MP.2021.2911}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yourname/vastR}
  \item Report bugs at \url{https://github.com/yourname/vastR/issues}
}

}
\author{
\strong{Maintainer}: Your Name \email{your.email@example.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

}
\keyword{internal}
